adding support for specifying rule import paths
