+++ stream_alert/rule_processor/main.py
@@ -14,45 +14,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 from __future__ import absolute_import  # Suppresses RuntimeWarning import error in Lambda
-import importlib
-import os
 
 from stream_alert.rule_processor.handler import StreamAlert
-
-
-def _python_rule_paths():
-    """Yields all .py files in matchers/ and rules/."""
-    # 'matchers' and 'rules' are top-level folders, both in the repo (when testing)
-    # and in the generated Lambda packages.
-    for folder in ('matchers', 'rules'):
-        for root, _, files in os.walk(folder):
-            for file_name in files:
-                if file_name.endswith('.py') and not file_name.startswith('__'):
-                    yield os.path.join(root, file_name)
-
-
-def _path_to_module(path):
-    """Convert a Python rules file path to an importable module name.
-
-    For example, "rules/community/cloudtrail_critical_api_calls.py" becomes
-    "rules.community.cloudtrail_critical_api_calls"
-
-    Raises:
-        NameError if a '.' appears anywhere in the path except the file extension.
-    """
-    base_name = os.path.splitext(path)[0]
-    if '.' in base_name:
-        raise NameError('Python file {} cannot be imported because of "." in the name', path)
-    return os.path.splitext(path)[0].replace('/', '.')
-
-
-def _import_rules():
-    """Dynamically import all rules files."""
-    for path in _python_rule_paths():
-        importlib.import_module(_path_to_module(path))
-
-
-_import_rules()
 
 
 def handler(event, context):
